Update files to match schema
