+++ packages/router/src/lib/entities/index.ts
@@ -14,3 +14,5 @@ export { PrepareInput, PrepareInputSchema } from "./prepare";
 export { FulfillInput, FulfillInputSchema } from "./fulfill";
 
 export { CancelInput, CancelInputSchema } from "./cancel";
+
+export * from "./metrics";
