+++ packages/router/src/lib/helpers/index.ts
@@ -10,3 +10,5 @@ export {
 export { getBidExpiry, AUCTION_EXPIRY_BUFFER } from "./auction";
 
 export { getNtpTimeSeconds } from "./shared";
+
+export { incrementFees, incrementGasConsumed, updateLiquidityGauge } from "./metrics";
