+++ packages/router/src/lib/helpers/metrics.ts
@@ -1,0 +1,7 @@
+import { BigNumber } from "@ethersproject/bignumber";
+
+export const incrementFees = async (assetId: string, chainId: number, amount: string) => {};
+
+export const incrementGasConsumed = async (chainId: number, gas: BigNumber) => {};
+
+export const updateLiquidityGauge = async (assetId: string, chainId: number, amount: string) => {};
